+++ src/features/calculator/components/dateTimePicker/StyledDateTimePicker.test.tsx
@@ -3,7 +3,9 @@ import userEvent from '@testing-library/user-event';
 import { describe, test, expect } from '@jest/globals';
 import { LocalizationProvider } from '@mui/x-date-pickers';
 import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
+import { Provider } from 'react-redux';
 import StyledDateTimePicker from './StyledDateTimePicker';
+import { store } from '../../../../app/store';
 
 describe('Test date time picker component', () => {
   configure({ testIdAttribute: 'data-test-id' });
@@ -12,9 +14,11 @@ describe('Test date time picker component', () => {
     test('Can save date and time input', async () => {
       const user = userEvent.setup();
       render(
-        <LocalizationProvider dateAdapter={AdapterDayjs}>
-          <StyledDateTimePicker />
-        </LocalizationProvider>,
+        <Provider store={store}>
+          <LocalizationProvider dateAdapter={AdapterDayjs}>
+            <StyledDateTimePicker />
+          </LocalizationProvider>
+        </Provider>,
       );
       const inputField = await screen.findByTestId('orderTime');
       await user.type(inputField, '010520241135');
@@ -24,9 +28,11 @@ describe('Test date time picker component', () => {
     test('Should report error when given time in the past', async () => {
       const user = userEvent.setup();
       render(
-        <LocalizationProvider dateAdapter={AdapterDayjs}>
-          <StyledDateTimePicker />
-        </LocalizationProvider>,
+        <Provider store={store}>
+          <LocalizationProvider dateAdapter={AdapterDayjs}>
+            <StyledDateTimePicker />
+          </LocalizationProvider>
+        </Provider>,
       );
       const inputField = await screen.findByTestId('orderTime');
       await user.type(inputField, '020120241135');